refactor: remove unused imports, console logs.
