+++ frontend/src/PrivateRoute.js
@@ -3,7 +3,6 @@ import { useLogin } from "./Reducer-context/LoginContext";
 
 export default function PrivateRoute({ path, ...props }) {
   const { isUserLogIn } = useLogin();
-  console.log("privateroute");
   return isUserLogIn ? (
     <Route {...props} path={path} />
   ) : (
