+++ frontend/src/pages/Fifteenpercent.jsx
@@ -1,8 +1,7 @@
 import { useEffect } from "react";
 import { useReduce } from "../Reducer-context/Reducer-context";
-import { Link } from "react-router-dom";
 import { Products } from "./Products";
-export function Tenpercent({
+export function Fifteenpercent({
   filteredData,
   Add_to_cart_button,
   Add_to_wishlist_button
@@ -11,7 +10,7 @@ export function Tenpercent({
   useEffect(() => {
     dispatch({ type: "MATERIAL", payload: "All" });
     dispatch({ type: "NEW", payload: false });
-    dispatch({ type: "OFFER", payload: 10 });
+    dispatch({ type: "OFFER", payload: 15 });
   }, []);
   return (
     <>
