+++ frontend/src/pages/New.jsx
@@ -1,8 +1,7 @@
 import { useEffect } from "react";
 import { useReduce } from "../Reducer-context/Reducer-context";
-import { Link } from "react-router-dom";
-import {Products} from "./Products";
-export function Seed({
+import { Products } from "./Products";
+export function New({
   filteredData,
   Add_to_cart_button,
   Add_to_wishlist_button
@@ -10,8 +9,8 @@ export function Seed({
   let { dispatch } = useReduce();
   useEffect(() => {
     dispatch({ type: "OFFER", payload: 0 });
-    dispatch({ type: "NEW", payload: false });
-    dispatch({ type: "MATERIAL", payload: "seed" });
+    dispatch({ type: "MATERIAL", payload: "All" });
+    dispatch({ type: "NEW", payload: true });
   }, []);
   return (
     <>
