+++ frontend/src/pages/Signup.jsx
@@ -1,31 +1,46 @@
 import { useLogin } from "../Reducer-context/LoginContext";
 import { useNavigate, useLocation, Link } from "react-router-dom";
 import { useEffect, useState } from "react";
+import axios from "axios";
 import treebanner from "../assets/treebanner.png";
 import { useReduce } from "../Reducer-context/Reducer-context";
-export function Login() {
-  let { isUserLogIn, LoginWithCredentials } = useLogin();
-  const { dispatch, user } = useReduce();
+
+export function Signup() {
+  let { isUserLogIn } = useLogin();
+  let { dispatch } = useReduce();
+  let [error, setError] = useState("");
+  const [userName, setUserName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   let navigate = useNavigate();
   let { state } = useLocation();
-  let [error, setError] = useState("");
+
   useEffect(() => {
     if (isUserLogIn) {
       navigate(state?.from ? state.from : "/", { replace: true });
     }
   }, [isUserLogIn]);
 
-  async function LoginHandler() {
-    dispatch({ type: "LOAD", payload: true });
-    let errorpassed = await LoginWithCredentials(email, password);
-    setError(errorpassed);
-    dispatch({ type: "LOAD", payload: false });
+  async function signupHandler() {
+    try {
+      dispatch({ type: "LOAD", payload: true });
+      let response = await axios.post(
+        "https://ecomm-demo-1.utpalpati.repl.co/signup",
+        { user: { userName, email, password } }
+      );
+      if (response.status === 200) {
+        setError("");
+        navigate("/login");
+      }
+      dispatch({ type: "LOAD", payload: false });
+    } catch (error) {
+      setError(error.response.data.message);
+      dispatch({ type: "LOAD", payload: false });
+    }
   }
   return (
-    <div className="login">
-      <div className="login-sideimg">
+    <div className="signup">
+      <div className="signup-sideimg">
         {" "}
         <img
           src={treebanner}
@@ -35,15 +50,26 @@ export function Login() {
       </div>
 
       <div className="form-container">
-        <h3>Login</h3>
+        <h3>Signup</h3>
 
         <form
           onSubmit={(e) => {
             e.preventDefault();
-            LoginHandler();
+            signupHandler();
           }}
           className="form"
         >
+          <label class="input md">
+            <input
+              type="text"
+              onChange={(e) => {
+                setUserName(e.target.value);
+              }}
+              class="input-text"
+              required
+            />
+            <span class="placeholder">Name</span>
+          </label>
           <label class="input md">
             <input
               type="email"
@@ -64,14 +90,14 @@ export function Login() {
               class="input-text"
               required
             />
-            <span class="placeholder">Password</span>
+            <span class="placeholder">password</span>
           </label>
           <div className="form-action">
             <button class="secondary-button md">
-              <Link to="/signup">Signup</Link>
+              <Link to="/login">Login</Link>
             </button>
             <button type="submit" class="secondary-button md">
-              Login
+              Signup
             </button>
           </div>
         </form>
