+++ frontend/src/pages/Tenpercent.jsx
@@ -1,8 +1,7 @@
 import { useEffect } from "react";
 import { useReduce } from "../Reducer-context/Reducer-context";
-import { Link } from "react-router-dom";
 import { Products } from "./Products";
-export function Fifteenpercent({
+export function Tenpercent({
   filteredData,
   Add_to_cart_button,
   Add_to_wishlist_button
@@ -11,7 +10,7 @@ export function Fifteenpercent({
   useEffect(() => {
     dispatch({ type: "MATERIAL", payload: "All" });
     dispatch({ type: "NEW", payload: false });
-    dispatch({ type: "OFFER", payload: 15 });
+    dispatch({ type: "OFFER", payload: 10 });
   }, []);
   return (
     <>
